Typing cleanup (#1150)
* add some # type ignore to text.__init__

* Fix DocumentLabel's document, it can not be None

* Fix all warning for text.formats.html

* fix some typing for image.__init__

* add isinstance check for __eq__

* some typing for text.__init__

* some typing fix for __init__

* add ignore for I001, change type: ignore to pyright: ignore

* now __init__ is clean!

* clean!

* ruff fix for lib

* cleanup for lib

* fixing whatever I found

* cleanup for event

* fix with cleanup

* remove unessarry isinstance check

* change image width/height define to a nicer look

* rework lib's noqa marker

* clean up for glx.py?

* clean up for wraptypes

* cleanups for gl

* cleanup for gengl

* clean up for advanced_font

* some more cleanup

* ignore some pyright for text

* add __all__ for window __init__

* remove _ViewEventHandler from window.__all__

* Fix example of advanced font

* format and add some ignore rule in pyproject

* FIX for examples

* add ignore for I001(import order), and add some more ignore for example

* fix lib

* Fix codeQL

* Fix pycharm typechecking in __init__

* more comment about advanced_font

* add a space before comment

* cleans for image

* font gdi -> float

* clean up some ANN

* cleanup for T201

* remove pathlib ignore

we should get start to merge old os.path code to pathlib, it's better, and old py are EOL, so, why not?

* cleanup gl

* Add back PTH123 for open()

* add exit in advanced_font example

* remove pyright ignore

* remove I001

* dpi into int

* remove #ruff

+++ pyglet/font/directwrite.py
@@ -2046,7 +2046,7 @@ class Win32DirectWriteFont(base.Font):
     texture_internalformat = pyglet.gl.GL_RGBA
 
     def __init__(self, name: str, size: float, bold: bool | str = False, italic: bool | str = False,
-                 stretch: bool | str = False, dpi: float | None = None, locale: str | None = None) -> None:
+                 stretch: bool | str = False, dpi: int | None = None, locale: str | None = None) -> None:
         self._filename: str | None = None
         self._advance_cache = {}  # Stores glyph's by the indice and advance.
