+++ pyglet/text/layout/incremental.py
@@ -106,7 +106,7 @@ class IncrementalTextLayout(TextLayout, EventDispatcher):
 
     def __init__(self, document: AbstractDocument, width: int, height: int, x: float = 0, y: float = 0, z: float = 0,  # noqa: D107
                  anchor_x: AnchorX = "left", anchor_y: AnchorY = "bottom", rotation: float = 0, multiline: bool = False,
-                 dpi: float | None = None, batch: Batch | None = None, group: Group | None = None,
+                 dpi: int | None = None, batch: Batch | None = None, group: Group | None = None,
                  program: ShaderProgram | None = None, wrap_lines: bool = True) -> None:
 
         if width is None or height is None:
